[SW-58214] habanalabs: change wait for interrupt timeout to 64 bit
In order to increase maximum wait for interrupt timeout, change it
to 64 bit variable.

+++ include/uapi/hlthunk.h
@@ -299,7 +299,7 @@ struct hlthunk_functions_pointers {
 	int (*fp_hlthunk_wait_for_interrupt)(int fd, void *addr,
 					uint64_t target_value,
 					uint32_t interrupt_id,
-					uint32_t timeout_us,
+					uint64_t timeout_us,
 					uint32_t *status);
 	int (*fp_hlthunk_command_submission_timeout)(int fd,
 						struct hlthunk_cs_in *in,
@@ -792,7 +792,7 @@ hlthunk_public int hlthunk_wait_for_multi_cs_with_timestamp(int fd,
 hlthunk_public int hlthunk_wait_for_interrupt(int fd, void *addr,
 					uint64_t target_value,
 					uint32_t interrupt_id,
-					uint32_t timeout_us,
+					uint64_t timeout_us,
 					uint32_t *status);
 
 /**
@@ -816,7 +816,7 @@ hlthunk_public int hlthunk_wait_for_interrupt(int fd, void *addr,
 hlthunk_public int hlthunk_wait_for_interrupt_with_timestamp(int fd, void *addr,
 					uint64_t target_value,
 					uint32_t interrupt_id,
-					uint32_t timeout_us,
+					uint64_t timeout_us,
 					uint32_t *status,
 					uint64_t *timestamp);
