+++ src/libhlthunk.c
@@ -1669,7 +1669,7 @@ hlthunk_public int hlthunk_wait_for_multi_cs_with_timestamp(int fd,
 lib_compat_public int hlthunk_wait_for_interrupt_v1_6(int fd, void *addr,
 					uint64_t target_value,
 					uint32_t interrupt_id,
-					uint32_t timeout_us,
+					uint64_t timeout_us,
 					uint32_t *status)
 {
 	return hlthunk_wait_for_interrupt_with_timestamp(
@@ -1692,14 +1692,14 @@ MAP_STATIC_SYMBOL(lib_compat_public int hlthunk_wait_for_interrupt(int fd,
 							void *addr,
 							uint64_t target_value,
 							uint32_t interrupt_id,
-							uint32_t timeout_us,
+							uint64_t timeout_us,
 							uint32_t *status),
 			hlthunk_wait_for_interrupt_v1_6);
 
 hlthunk_public int hlthunk_wait_for_interrupt_with_timestamp(int fd, void *addr,
 					uint64_t target_value,
 					uint32_t interrupt_id,
-					uint32_t timeout_us,
+					uint64_t timeout_us,
 					uint32_t *status,
 					uint64_t *timestamp)
 {
